+++ snppipeline/snppipeline.py
@@ -118,22 +118,22 @@ def create_snp_list(options_dict):
     #==========================================================================
     # Prep work
     #==========================================================================
-    sample_directories_list_filename = options_dict['sampleDirsFile']
-    bad_file_count = utils.verify_non_empty_input_files("File of sample directories", [sample_directories_list_filename])
+    sample_directories_list_path = options_dict['sampleDirsFile']
+    bad_file_count = utils.verify_non_empty_input_files("File of sample directories", [sample_directories_list_path])
     if bad_file_count > 0:
         utils.global_error(None)
 
-    with open(sample_directories_list_filename, "r") as sample_directories_list_file:
-        list_of_sample_directories = [line.rstrip() for line in sample_directories_list_file]
-    list_of_sample_directories = sorted([d for d in list_of_sample_directories if d])
+    with open(sample_directories_list_path, "r") as sample_directories_list_file:
+        unsorted_list_of_sample_directories = [line.rstrip() for line in sample_directories_list_file]
+    unsorted_list_of_sample_directories = [d for d in unsorted_list_of_sample_directories if d]
+    sorted_list_of_sample_directories = sorted(unsorted_list_of_sample_directories)
 
     #==========================================================================
-    # Read in all vcf files and process into dict of SNPs passing various
-    # criteria. Do this for each sample. Write to file.
+    # Validate inputs
     #==========================================================================
     snp_list_file_path = options_dict['snpListFile']
     vcf_file_name = options_dict['vcfFileName']
-    list_of_vcf_files = [os.path.join(dir, vcf_file_name) for dir in list_of_sample_directories]
+    list_of_vcf_files = [os.path.join(dir, vcf_file_name) for dir in sorted_list_of_sample_directories]
 
     bad_file_count = utils.verify_non_empty_input_files("VCF file", list_of_vcf_files)
     if bad_file_count == len(list_of_vcf_files):
@@ -141,11 +141,51 @@ def create_snp_list(options_dict):
     elif bad_file_count > 0:
         utils.sample_error("Error: %d VCF files were missing or empty." % bad_file_count, continue_possible=True)
 
+    #==========================================================================
+    # Read in all vcf files and process into dict of SNPs passing various
+    # criteria. Do this for each sample. Write to file.
+    #==========================================================================
     if options_dict['forceFlag'] or utils.target_needs_rebuild(list_of_vcf_files, snp_list_file_path):
-        snp_dict = utils.convert_vcf_files_to_snp_dict(list_of_vcf_files)
+        snp_dict = dict()
+        excluded_sample_directories = set()
+        for sample_dir, vcf_file_path in zip(sorted_list_of_sample_directories, list_of_vcf_files):
+
+            if not os.path.isfile(vcf_file_path):
+                continue
+            if os.path.getsize(vcf_file_path) == 0:
+                continue
+
+            verbose_print("Processing VCF file %s" % vcf_file_path)
+            sample_name = os.path.basename(os.path.dirname(vcf_file_path))
+            snp_set = utils.convert_vcf_file_to_snp_set(vcf_file_path)
+            max_snps = options_dict['maxSnps']
+            if max_snps >= 0 and len(snp_set) > max_snps:
+                verbose_print("Excluding sample %s having %d snps." % (sample_name, len(snp_set)))
+                excluded_sample_directories.add(sample_dir)
+                continue
+    
+            for key in snp_set:
+                if key not in snp_dict:
+                    sample_list = [sample_name]
+                    snp_dict[key] = sample_list
+                else:
+                    sample_list = snp_dict[key]
+                    sample_list.append(sample_name)
+
         verbose_print('Found %d snp positions across %d sample vcf files.' % (len(snp_dict), len(list_of_vcf_files)))
         utils.write_list_of_snps(snp_list_file_path, snp_dict)
         verbose_print("")
+
+        #==========================================================================
+        # Write the filtered list of sample directories
+        #==========================================================================
+        sample_directories_list_path = sample_directories_list_path + ".filtered"
+        with open(sample_directories_list_path, "w") as filtered_samples_file_object:
+            # Loop over the unsorted list to keep the order of samples the same as the original.
+            # This will keep the same HPC log file suffix number.
+            for sample_dir in unsorted_list_of_sample_directories:
+                if sample_dir not in excluded_sample_directories:
+                    filtered_samples_file_object.write("%s\n" % sample_dir)
     else:
         verbose_print("SNP list %s has already been freshly built.  Use the -f option to force a rebuild." % snp_list_file_path)
     verbose_print("# %s %s finished" % (utils.timestamp(), utils.program_name()))